+++ views/admin-page-settings.php
@@ -14,7 +14,13 @@ if ( ! defined( 'ABSPATH' ) ) {
 <div class="wrap prpl-wrap prpl-settings-wrap">
 	<div class="prpl-header">
 		<div class="prpl-header-logo">
-			<?php \progress_planner()->the_asset( 'images/logo_progress_planner.svg' ); ?>
+			<?php
+			if ( \progress_planner()->is_pro_site() ) {
+				\progress_planner()->the_asset( 'images/logo_progress_planner_pro.svg' );
+			} else {
+				\progress_planner()->the_asset( 'images/logo_progress_planner.svg' );
+			}
+			?>
 		</div>
 	</div>
 	<h1>