+++ views/page-widgets/activity-scores.php
@@ -56,7 +56,14 @@ $prpl_record = $prpl_widget->personal_record_callback();
 	\progress_planner()->get_chart()->the_chart(
 		[
 			'type'           => 'bar',
-			'query_params'   => [],
+			'items_callback' => function ( $start_date, $end_date ) {
+				return \progress_planner()->get_query()->query_activities(
+					[
+						'start_date' => $start_date,
+						'end_date'   => $end_date,
+					]
+				);
+			},
 			'dates_params'   => [
 				'start_date' => \DateTime::createFromFormat( 'Y-m-d', \gmdate( 'Y-m-01' ) )->modify( $prpl_widget->get_range() ),
 				'end_date'   => new \DateTime(),